HUDTextの作成

+++ KIFoundation/TriangleIterator.cpp
@@ -2,7 +2,7 @@ namespace KI
 {
 namespace Foundation
 {
-TriangleIterator::TriangleIterator(unsigned int sourceType, const vec3* position, const int* indexs, int size)
+TriangleIterator::TriangleIterator(unsigned int sourceType, const vec3* position, const int* indexs, size_t size)
 {
 	if (sourceType == GL_TRIANGLES) {
 		m_sourceType = TriangleType::TRIANGLES;
@@ -16,7 +16,7 @@ TriangleIterator::TriangleIterator(unsigned int sourceType, const vec3* position
 	m_size = size;
 }
 
-TriangleIterator::TriangleIterator(TriangleType sourceType, const vec3* position, const int* indexs, int size)
+TriangleIterator::TriangleIterator(TriangleType sourceType, const vec3* position, const int* indexs, size_t size)
 {
 	m_sourceType = sourceType;
 	m_position = position;
