+++ RenderModel/PlotUI.cpp
@@ -4,13 +4,13 @@ namespace ImGuiEx
 {
 bool PlotLineUI::Show(std::vector<float>* value)
 {
-	ImGui::PlotLines(m_label.data(), value->data(), value->size());
+	ImGui::PlotLines(m_label.data(), value->data(), (int)value->size());
 	return true;
 }
 
 bool PlotLineUI::Show(const std::vector<float>& value)
 {
-	ImGui::PlotLines(m_label.data(), value.data(), value.size());
+	ImGui::PlotLines(m_label.data(), value.data(), (int)value.size());
 	return true;
 }
 }
